Add cli executable test
