+++ tests_integration/runPrettier.js
@@ -4,15 +4,7 @@ const fs = require("fs");
 const path = require("path");
 const stripAnsi = require("strip-ansi");
 const SynchronousPromise = require("synchronous-promise").SynchronousPromise;
-
-const isProduction = process.env.NODE_ENV === "production";
-const prettierRootDir = isProduction ? process.env.PRETTIER_DIR : "../";
-const prettierPkg = require(path.join(prettierRootDir, "package.json"));
-const prettierCli = path.join(prettierRootDir, prettierPkg.bin.prettier);
-
-const thirdParty = isProduction
-  ? path.join(prettierRootDir, "./third-party")
-  : path.join(prettierRootDir, "./src/common/third-party");
+const { thirdParty, prettierCli } = require("./env");
 
 function runPrettier(dir, args, options) {
   args = args || [];
